SIM: drm/i915: Add the simulator placehoder and parameter.
Usually the simulator detection can be done according to HAS PCH
but if we are in a platform without PCH or if we force a different
PCI ID than simulator is expecting that auto detection fails.

So, let's introduce the placehoder but also include a way to force
is_simulator during boot.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>