FIXME: drm/i915/tgl: tgl 12 s/ss/eu fuse reading support
Same registers as Gen11, with the difference that in the sub-slice fuse
1 now means enabled. To avoid confusion, a new definition for the
register has been added for Gen12.
Also, tgl only supports slice-level power gating.

FIXME: tgl has fused EUs, where a "big" EU is paired with a "little"
EU. The big EUs can do all operations, while the small ones only 3D and
GPGPU (i.e. no media). The maximum number of available EUs is 8 pairs
(16 total EUs) per subslice. Only 8 bits in the fuse register are
defined (Bspec: 9731), so I've assumed that the value refers to EU
pairs (I can't find any details on this in either Bspec or HSDs). We
also need to decide if we want to report the number of pairs or the
number of total EUs to UMDs for thread count calculations. The
expectation is that in most cases media workload will only use the big
EUs while 3D and GPGPU will use all of them (EU kernel support for fused
EUs is required). For simplicitly, this patch keeps the reporting to
number of pairs as that's what seems to come out of the fuse register
(with the added benefit that the patch is smaller and easier to rebase).

v2 (James): Rebase

Bspec: 29547
Bspec: 12247
Cc: Michel Thierry <michel.thierry@intel.com>
Cc: Kelvin Gardiner <kelvin.gardiner@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: James Ausmus <james.ausmus@intel.com>