drm/i915/ICL: Add register definition for DFLEXDPMLE
DFLEXDPMLE register is required to tell the FIA hardware which
main links of DP are enabled on Type-C Connectors. FIA uses this
information to program PHY to Controller signal mapping.
This register is applicable in both Type C connector's Alternate mode
as well as DP connector mode.

v3:
* Define one mask and bit definition and take a connector number
as input (Jani Nikula)
v2:
* Add two extra spaces for bit and mask defs (Madhav)
* Define most significant bits first (Madhav)

Cc: Jani Nikula <jani.nikula@linux.intel.com>
Cc: Animesh Manna <animesh.manna@intel.com>
Cc: Madhav Chauhan <madhav.chauhan@intel.com>
Cc: Anusha Srivatsa <anusha.srivatsa@intel.com>
Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Manasi Navare <manasi.d.navare@intel.com>