drm/i915/icl: WaForceEnableNonCoherent
This is not the same workaround as in early Gen9 platforms, where
lacking this could cause system hangs. But coherency performance
overhead is high and only a few compute workloads really need it
(the register is whitelisted in hardware, so UMDs can opt in for
coherency if they have a good reason).

Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>