drm/i915/tgl: Enable ccs/dual-ctx RCU and GAM registers and reduce ppgtt size
We have to specify in the Render Control Unit Mode and GAM MultiContext
Control registers when CCS is enabled.

These bits tell the hardware to use the MSB bit to split between RCS and
CCS when dual context execution is happening. So the only virtual address
mode supported is 48-bit PPGTT and with 47-bit size.

v2: Combine register setup and size reduction in the same commit.

Bspec: 29634, 29686, 28700.
Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com> (v1)