drm/i915/icl/guc: Add support for the new GuC workqueue item submission mechanism
To simplify the scheduling logic in the GuC firmware, starting Gen11
only out-of-order mode of scheduling is supported. This basically means
that the KMD is now in charge of updating the tail inside the context
before submitting to the GuC backend (so the GuC will not track individual
requests anymore, only contexts).

Currently, in-orderness is guaranteed by the execlists emulation with two
"fake" ports that the scheduler uses when submitting to the GuC backend.
If we are submitting requests for context A in the first port and we place
a request for context B in the second port, we won't submit more requests
for A until all the pending ones complete.

We have to take this into account if we try to change the current execlist
emulation model (e.g.: increasing the number of fake ports could cause
requests to execute in wrong global seqno order).

v2: Rebased
v3: Rebased
v4: Non-trivial rebase. After guc_wq_item_append function split,
    guc_wq_item_append no longer receives a request object (because
    preemption doesn't use one). The problem is preGen11 is ok updating the
    ring tail, but in Gen11 we want to update the ring tail in the lrc context
    instead. Code changed to pass enough information to just update the ring tail
    in the lrc_reg_state, instead of calling the function that updates the whole
    context. Note: Added a GEM_BUG_ON to break preemption in Gen11+ because the
    GuC command is different there anyway. (Michel).
v5: Less parameters required to guc_wq_item_append (Oscar)
v6: Rebased

Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Michał Winiarski <michal.winiarski@intel.com>