SIM: drm/i915: ICL+ needs fake interrupts for pipes
Although pipe A interrupts work, we still need the fake code por pipes
B and C.

This not only removes the error messages, but it also makes the boot
on the simulator a lot faster.

v2: Upstream moved and we don't get more flip inteerrupts, so we need
for pipe A too, unless we want we want to subscribe to flip
interrupts, which appear to work on pipe A.

v3 (Daniele): extend to TGL, update commit title, add BUG_ONs to ensure
we don't use the wrong version of the function.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>