drm/i915/icl: add basic support for the ICL clocks
This commit introduces the definitions for the ICL clocks and adds the
basic functions to the shared DPLL framework. It adds code for the
Enable and Disable sequences for the Combo, TBT and MG PLLs, but it
does not have the code to compute the actual PLL values, which are
marked as TODO comments and should be introduced as separate commits.

v2 (James): Remove grab of dpll_lock inside IS_ICELAKE block of
intel_ddi_clk_select, as the mutex is grabbed at the top of the function
as of upstream '8edcda1266f9 ("drm/i915: Protect DDI port to DPLL map
from theoretical race.")'

v3: fix dpll id comments (Madhav, Vandita)

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: James Ausmus <james.ausmus@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>